requirements.txt
asgiref==3.5.0
Django==3.2.12
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.10.1
Markdown==3.3.6
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
zipp==3.7.0
names~=0.3.0
mixer~=7.2.1